% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_age.R
\name{say_age}
\alias{say_age}
\title{Say age!}
\usage{
say_age(my_name, my_age, punctuation)
}
\arguments{
\item{my_name}{user's name}

\item{my_age}{user's age}

\item{punctuation}{any punctuation the user wants to end the sentence with}
}
\description{
A function to report a user's name and age
}
\examples{
say_age("Liz", 21, "!")
}
\author{
Elizabeth Grace Peterson
}
